# aspnetcore/src/IO.Swagger/Controllers/InventoryAssetsApi.cs
/*
 * Digitalization of Asset Counting Web Api
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class InventoryAssetsApiController : ControllerBase
    { 
        /// <summary>
        /// Updates an InventoryOrderAsset instance.
        /// </summary>
        
        /// <param name="inventoryOrderAsset">The inventoryOrderAsset object to be updated.</param>
        /// <response code="200">Success</response>
        [HttpPost]
        [Route("/api/inventoryassets")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("Post")]
        [SwaggerResponse(statusCode: 200, type: typeof(int?), description: "Success")]
        public virtual IActionResult Post([FromBody]InventoryOrderAsset inventoryOrderAsset)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(int?));

            string exampleJson = null;
            exampleJson = "0";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<int?>(exampleJson)
            : default(int?);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Updates an InventoryOrderAsset instance.
        /// </summary>
        
        /// <param name="inventoryOrderAssets">The inventoryOrderAsset list to be updated.</param>
        /// <response code="200">Success</response>
        [HttpPost]
        [Route("/api/inventoryassets/bulk")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("Post_0")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<int?>), description: "Success")]
        public virtual IActionResult Post_0([FromBody]List<InventoryOrderAsset> inventoryOrderAssets)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<int?>));

            string exampleJson = null;
            exampleJson = "[ 0, 0 ]";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<List<int?>>(exampleJson)
            : default(List<int?>);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}