# aspnetcore/src/IO.Swagger/Controllers/SapInventoryAssetVariantsUpdaterApi.cs
/*
 * Digitalization of Asset Counting Web Api
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class SapInventoryAssetVariantsUpdaterApiController : ControllerBase
    { 
        /// <summary>
        /// Update the permissions after new import of assets or import of inventory variants.
        /// </summary>
        
        /// <param name="sourceSystem">the sap source system that caused the change in data</param>
        /// <response code="204">Success</response>
        [HttpGet]
        [Route("/api/sapinventoryassets/variants/update")]
        [Authorize(AuthenticationSchemes = ApiKeyAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("Get")]
        public virtual IActionResult Get([FromQuery]string sourceSystem)
        { 
            //TODO: Uncomment the next line to return response 204 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(204);


            throw new NotImplementedException();
        }
    }
}