# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v1",
    "title" : "Digitalization of Asset Counting Web Api"
  },
  "security" : [ {
    "Bearer" : [ ],
    "Basic" : [ ]
  } ],
  "paths" : {
    "/api/appversions" : {
      "get" : {
        "tags" : [ "AppVersions" ],
        "summary" : "Returns current iphone App Version.",
        "operationId" : "Get",
        "consumes" : [ ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "string"
            }
          }
        },
        "deprecated" : true
      },
      "post" : {
        "tags" : [ "AppVersions" ],
        "summary" : "set current iphone App Version.",
        "operationId" : "Post",
        "consumes" : [ ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "name" : "version",
          "in" : "query",
          "description" : "Update app current version to this value.",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        },
        "deprecated" : true
      }
    },
    "/api/v2/appversions" : {
      "get" : {
        "tags" : [ "AppVersionsControllerV2" ],
        "summary" : "Returns current iphone App Version.",
        "operationId" : "Get",
        "consumes" : [ ],
        "produces" : [ ],
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        }
      },
      "post" : {
        "tags" : [ "AppVersionsControllerV2" ],
        "summary" : "set current iphone App Version.",
        "operationId" : "Post",
        "consumes" : [ ],
        "produces" : [ ],
        "parameters" : [ {
          "name" : "version",
          "in" : "query",
          "description" : "Update app current version to this value.",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        }
      }
    },
...
    "/api/inventories" : {
      "get" : {
        "tags" : [ "Inventories" ],
        "summary" : "Returns list of InventoryOrders. This query can be paged. If page and pageSize are omitted,\r\nthe full list is going to be returned. When pageSize and page specified,\r\npageCount and currentPage will be returned on first call.",
        "operationId" : "Get",
        "consumes" : [ ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "description" : "define the page that this query will return",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "pageSize",
          "in" : "query",
          "description" : "define the number of items each page will contain if records are available",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/InventoryOrder"
              },
              "uniqueItems" : false
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Inventories" ],
        "summary" : "Updates an InventoryOrder instance.",
        "operationId" : "Post",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "in" : "body",
          "name" : "inventoryOrder",
          "description" : "The inventoryOrder object to be updated.",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/InventoryOrder"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "deprecated" : true
      },
      "put" : {
        "tags" : [ "Inventories" ],
        "summary" : "Creates an InventoryOrder instance.",
        "operationId" : "Put",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "in" : "body",
          "name" : "inventoryOrderCreation",
          "description" : "Defines the inventory order to be created and the inventory variant that is used to populate assets.",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/InventoryOrderCreation"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "deprecated" : true
      }
    },
...
    "/api/v2/inventories" : {
      "post" : {
        "tags" : [ "InventoriesControllerV2" ],
        "summary" : "Creates an InventoryOrder instance.",
        "operationId" : "Create",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "inventoryOrderCreation",
          "description" : "Defines the inventory order to be created and the inventory variant that is used to populate assets.",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/InventoryOrderCreation"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/InventoryOrder"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "InventoriesControllerV2" ],
        "summary" : "Updates InventoryOrder instances.",
        "operationId" : "Updates",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ ],
        "parameters" : [ {
          "in" : "body",
          "name" : "inventoryOrders",
          "description" : "The inventoryOrders to be updated.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/InventoryOrderPutResource"
            },
            "uniqueItems" : false
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        }
      }
    },
...
    "/api/inventoryassets" : {
      "post" : {
        "tags" : [ "InventoryAssets" ],
        "summary" : "Updates an InventoryOrderAsset instance.",
        "operationId" : "Post",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "inventoryOrderAsset",
          "description" : "The inventoryOrderAsset object to be updated.",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/InventoryOrderAsset"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        },
        "deprecated" : true
      }
    },
    "/api/inventoryassets/bulk" : {
      "post" : {
        "tags" : [ "InventoryAssets" ],
        "summary" : "Updates an InventoryOrderAsset instance.",
        "operationId" : "Post",
        "consumes" : [ "application/json-patch+json", "application/json", "text/json", "application/*+json" ],
        "produces" : [ "text/plain", "application/json", "text/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "inventoryOrderAssets",
          "description" : "The inventoryOrderAsset list to be updated.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/definitions/InventoryOrderAsset"
            },
            "uniqueItems" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "integer",
                "format" : "int32"
              },
              "uniqueItems" : false
            }
          }
        },
        "deprecated" : true
      }
    },
...
    "/api/sapinventoryassets/variants/update" : {
      "get" : {
        "tags" : [ "SapInventoryAssetVariantsUpdater" ],
        "summary" : "Update the permissions after new import of assets or import of inventory variants.",
        "operationId" : "Get",
        "consumes" : [ ],
        "produces" : [ ],
        "parameters" : [ {
          "name" : "sourceSystem",
          "in" : "query",
          "description" : "the sap source system that caused the change in data",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "Bearer" : {
      "description" : "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "Asset" : {
      "type" : "object",
      "required" : [ "assetNumber", "companyCode", "subNumber" ],
      "properties" : {
        "tagId" : {
          "type" : "string",
          "maxLength" : 24
        },
        "secondTagId" : {
          "type" : "string",
          "maxLength" : 24
        },
        "isUntaggable" : {
          "type" : "boolean"
        },
        "latitude" : {
          "type" : "number",
          "format" : "double"
        },
        "longitude" : {
          "type" : "number",
          "format" : "double"
        },
        "lastModified" : {
          "type" : "string",
          "format" : "date-time"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "creator" : {
          "type" : "string",
          "maxLength" : 255
        },
        "comment" : {
          "type" : "string",
          "maxLength" : 254
        },
        "shareTags" : {
          "type" : "boolean"
        },
        "sapAsset" : {
          "$ref" : "#/definitions/SapInventoryAsset"
        },
        "assetNumber" : {
          "type" : "string",
          "maxLength" : 12
        },
        "subNumber" : {
          "type" : "string",
          "maxLength" : 12
        },
        "companyCode" : {
          "type" : "string",
          "maxLength" : 12
        }
      },
      "example" : {
        "companyCode" : "companyCode",
        "creator" : "creator",
        "sapAsset" : {
          "companyCode" : "companyCode",
          "serialNumber" : "serialNumber",
          "quantity" : 5.962133916683182,
          "costCenter" : "costCenter",
          "capitalisationDate" : "2000-01-23T04:56:07.000+00:00",
          "deactivatedOn" : "2000-01-23T04:56:07.000+00:00",
          "typeName" : "typeName",
          "description" : "description",
          "acquisitionValue" : 1.4658129805029452,
          "assetClass" : "assetClass",
          "room" : "room",
          "subNumber" : "subNumber",
          "inventoryNumber" : "inventoryNumber",
          "unit" : "unit",
          "assetNumber" : "assetNumber",
          "isDeactivated" : true,
          "vendor" : "vendor",
          "plant" : "plant",
          "location" : "location",
          "assetSuperNumber" : "assetSuperNumber",
          "lastInventoryDate" : "2000-01-23T04:56:07.000+00:00",
          "companyCodeCurrency" : "companyCodeCurrency"
        },
        "tagId" : "tagId",
        "created" : "2000-01-23T04:56:07.000+00:00",
        "latitude" : 0.8008281904610115,
        "shareTags" : true,
        "isUntaggable" : true,
        "subNumber" : "subNumber",
        "secondTagId" : "secondTagId",
        "assetNumber" : "assetNumber",
        "comment" : "comment",
        "lastModified" : "2000-01-23T04:56:07.000+00:00",
        "longitude" : 6.027456183070403
      }
    },
...
    "InventoryOrder" : {
      "type" : "object",
      "required" : [ "id" ],
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "description" : {
          "type" : "string",
          "maxLength" : 255
        },
        "inventoryVariantName" : {
          "type" : "string",
          "maxLength" : 255
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "reportDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "creator" : {
          "type" : "string",
          "maxLength" : 255
        },
        "assignees" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/InventoryOrderAssignee"
          }
        },
        "status" : {
          "type" : "integer",
          "format" : "int32",
          "enum" : [ 0, 1, 2, 3, 4 ]
        },
        "hasOnlyUntaggedAssets" : {
          "type" : "boolean"
        }
      },
      "example" : {
        "creator" : "creator",
        "reportDate" : "2000-01-23T04:56:07.000+00:00",
        "created" : "2000-01-23T04:56:07.000+00:00",
        "hasOnlyUntaggedAssets" : true,
        "inventoryVariantName" : "inventoryVariantName",
        "description" : "description",
        "assignees" : [ {
          "id" : "id",
          "assignee" : "assignee",
          "inventoryOrderId" : 6
        }, {
          "id" : "id",
          "assignee" : "assignee",
          "inventoryOrderId" : 6
        } ],
        "id" : 0,
        "status" : 1
      }
    },
    "InventoryOrderAssignee" : {
      "type" : "object",
      "required" : [ "assignee", "id", "inventoryOrderId" ],
      "properties" : {
        "inventoryOrderId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "id" : {
          "type" : "string",
          "maxLength" : 24
        },
        "assignee" : {
          "type" : "string",
          "maxLength" : 255
        }
      },
      "example" : {
        "id" : "id",
        "assignee" : "assignee",
        "inventoryOrderId" : 6
      }
    },
    "InventoryOrderCreation" : {
      "type" : "object",
      "properties" : {
        "hasOnlyUntaggedAssets" : {
          "type" : "boolean"
        },
        "inventoryVariantName" : {
          "type" : "string"
        },
        "reportDate" : {
          "type" : "string",
          "format" : "date-time"
        }
      },
      "example" : {
        "reportDate" : "2000-01-23T04:56:07.000+00:00",
        "hasOnlyUntaggedAssets" : true,
        "inventoryVariantName" : "inventoryVariantName"
      }
    },
    "InventoryOrderPutResource" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "status" : {
          "type" : "integer",
          "format" : "int32",
          "enum" : [ 0, 1, 2, 3, 4 ]
        }
      }
    },
...
    "InventoryOrderAsset" : {
      "type" : "object",
      "required" : [ "assetNumber", "companyCode", "inventoryOrderId", "subNumber" ],
      "properties" : {
        "inventoryOrderId" : {
          "type" : "integer",
          "format" : "int32"
        },
        "checked" : {
          "type" : "boolean"
        },
        "lastModified" : {
          "type" : "string",
          "format" : "date-time"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "changedBy" : {
          "type" : "string",
          "maxLength" : 254
        },
        "commentOther" : {
          "type" : "string",
          "maxLength" : 254
        },
        "commentCostCenter" : {
          "type" : "string",
          "maxLength" : 254
        },
        "commentPlant" : {
          "type" : "string",
          "maxLength" : 254
        },
        "commentLocation" : {
          "type" : "string",
          "maxLength" : 254
        },
        "status" : {
          "type" : "integer",
          "format" : "int32",
          "enum" : [ 0, 1, 2, 3, 4, 5 ]
        },
        "asset" : {
          "$ref" : "#/definitions/Asset"
        },
        "assetNumber" : {
          "type" : "string",
          "maxLength" : 12
        },
        "subNumber" : {
          "type" : "string",
          "maxLength" : 12
        },
        "companyCode" : {
          "type" : "string",
          "maxLength" : 12
        }
      },
      "example" : {
        "companyCode" : "companyCode",
        "created" : "2000-01-23T04:56:07.000+00:00",
        "commentCostCenter" : "commentCostCenter",
        "inventoryOrderId" : 0,
        "subNumber" : "subNumber",
        "commentLocation" : "commentLocation",
        "assetNumber" : "assetNumber",
        "changedBy" : "changedBy",
        "checked" : true,
        "commentPlant" : "commentPlant",
        "lastModified" : "2000-01-23T04:56:07.000+00:00",
        "commentOther" : "commentOther",
        "asset" : {
          "companyCode" : "companyCode",
          "creator" : "creator",
          "sapAsset" : {
            "companyCode" : "companyCode",
            "serialNumber" : "serialNumber",
            "quantity" : 5.962133916683182,
            "costCenter" : "costCenter",
            "capitalisationDate" : "2000-01-23T04:56:07.000+00:00",
            "deactivatedOn" : "2000-01-23T04:56:07.000+00:00",
            "typeName" : "typeName",
            "description" : "description",
            "acquisitionValue" : 1.4658129805029452,
            "assetClass" : "assetClass",
            "room" : "room",
            "subNumber" : "subNumber",
            "inventoryNumber" : "inventoryNumber",
            "unit" : "unit",
            "assetNumber" : "assetNumber",
            "isDeactivated" : true,
            "vendor" : "vendor",
            "plant" : "plant",
            "location" : "location",
            "assetSuperNumber" : "assetSuperNumber",
            "lastInventoryDate" : "2000-01-23T04:56:07.000+00:00",
            "companyCodeCurrency" : "companyCodeCurrency"
          },
          "tagId" : "tagId",
          "created" : "2000-01-23T04:56:07.000+00:00",
          "latitude" : 0.8008281904610115,
          "shareTags" : true,
          "isUntaggable" : true,
          "subNumber" : "subNumber",
          "secondTagId" : "secondTagId",
          "assetNumber" : "assetNumber",
          "comment" : "comment",
          "lastModified" : "2000-01-23T04:56:07.000+00:00",
          "longitude" : 6.027456183070403
        },
        "status" : 6
      }
    },
...
  }
}